Render free credits modal at top level
