+++ src/components/NewWorkspaceModal.js
@@ -7,7 +7,7 @@ import { icon } from 'src/components/icons'
 import { TextArea, validatedInput } from 'src/components/input'
 import Modal from 'src/components/Modal'
 import { InfoBox } from 'src/components/PopupTrigger'
-import { freeCreditsActive } from 'src/components/TrialBanner'
+import { freeCreditsActive } from 'src/components/FreeCreditsModal'
 import { ajaxCaller } from 'src/libs/ajax'
 import { authStore } from 'src/libs/auth'
 import colors from 'src/libs/colors'
