+++ src/pages/Main.js
@@ -3,6 +3,7 @@ import { hot } from 'react-hot-loader/root'
 import { h } from 'react-hyperscript-helpers'
 import ConfigOverridesWarning from 'src/components/ConfigOverridesWarning'
 import ErrorWrapper from 'src/components/ErrorWrapper'
+import FreeCreditsModal from 'src/components/FreeCreditsModal'
 import Notifications from 'src/components/Notifications'
 import { NpsSurvey } from 'src/components/NpsSurvey'
 import Router from 'src/components/Router'
@@ -13,6 +14,7 @@ import { TrialBanner } from 'src/components/TrialBanner'
 const Main = () => {
   return h(Fragment, [
     h(Notifications),
+    h(FreeCreditsModal),
     h(ErrorWrapper, [h(TrialBanner), h(Router)]),
     h(SupportRequest),
     h(NpsSurvey),
